requirements.txt
# 核心依赖
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0

# Git 管理
gitpython==3.1.42

# 向量数据库
chromadb==0.4.22
sentence-transformers==2.3.1

# HTTP 客户端
httpx==0.26.0
aiohttp==3.9.3

# 工具库
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
pyyaml==6.0.1

# 日志
loguru==0.7.2

# 测试
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0

# 开发工具
black==24.2.0
isort==5.13.2
mypy==1.8.0